Basic UI (#4)
* Update main.py

* Line Numbers to left text box

* Line Numbers More Responsive

* Another Teeny but big update

* Tabs UI Update

* Update main.py

* Tabs part 2

* Major B''H Update

* Blah dee blah

* Tabs number almost done

* Update main.py

update text in view box

* Update main.py

* Fixes Line Numbers

* Updated Tab Management

(1) Modified the code to disable the delete button (X) when only one tab is left and enable it when there are multiple tabs.

(2) Ensured that when a tab is closed, the program automatically 'clicks' on one of the remaining tabs to keep an active tab.

* Open File now works with tabs

* Somewhat Rename Tab

* BUG FIX: Open file Tab Switching Error

* Save file renames tab

* Takes out debug print statements

* Bug Fix

Only can save one file. Now the user can save multiple files.

* Icon and Markdown Box selectbackground

* Closing and Opening Tabs Redraws Line Numbers

This is so that it shows the correct line numbers

---------

Co-authored-by: Professor Howard Jack Yont <dovidstahler9@gmail.com>